requirements.txt
alembic==1.7.1
attrs==21.2.0
click==8.0.1
Flask==1.1.2
Flask-Cors==3.0.10
flask-expects-json==1.6.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.25.0
greenlet==1.1.1
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
mysql-client==0.0.1
mysqlclient==2.0.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyrsistent==0.18.0
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
